build(header): header 파일 pages/header로 이동
