+++ frontend/main.js
@@ -1,5 +1,5 @@
 import App from "./src/app.js";
-import Header from "./src/common/header.js";
+import Header from "./src/pages/header/header.js";
 import { $ } from "./src/utils/querySelector.js";
 
 window.addEventListener("DOMContentLoaded", (e) => {
